pyproject.toml
[tool.poetry]
name = "py-home-music-downloader"
version = "1.0.0"
description = ""
authors = ["seanbranner <seanbranner@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
coverage = "5.5"
deprecation = "2.1.0"
ffmpeg = "1.4"
filetype = "1.0.8"
mutagen = "1.45.1"
packaging = "21.0"
pydub = "0.25.1"
pyparsing = "2.4.7"
toml = "0.10.2"
python-dotenv = "^1.0.1"
youtube-dl = "^2021.12.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
